requirements.txt
matplotlib
networkx
pickle
python-telegram-bot
antlr4-python3-runtime